migrate $ref sibling validator to 2and3
