+++ src/plugins/validate-semantic/index.js
@@ -12,8 +12,8 @@ import * as securityValidateActions from "./validators/security"
 import * as refsValidateActions from "./validators/refs"
 import * as parametersValidateActions from "./validators/parameters"
 import * as operationsValidateActions from "./validators/operations"
-import * as refsOAS3ValidateActions from "./validators/oas3/refs"
 import * as operationsOAS3ValidateActions from "./validators/oas3/operations"
+import * as refs2and3ValidateActions from "./validators/2and3/refs"
 import * as parameters2and3ValidateActions from "./validators/2and3/parameters"
 import * as paths2and3ValidateActions from "./validators/2and3/paths"
 
@@ -54,7 +54,7 @@ export default function SemanticValidatorsPlugin({getSystem}) {
           ...refsValidateActions,
           ...parametersValidateActions,
           ...operationsValidateActions,
-          ...refsOAS3ValidateActions,
+          ...refs2and3ValidateActions,
           ...operationsOAS3ValidateActions,
           ...parameters2and3ValidateActions,
           ...paths2and3ValidateActions
