+++ test/plugins/validate-semantic/refs.js
@@ -3,35 +3,6 @@ import validateHelper, { expectNoErrorsOrWarnings } from "./validate-helper.js"
 
 describe("validation plugin - semantic - refs", function() {
   this.timeout(10 * 1000)
-  describe("Ref siblings", () => {
-
-    it("should return a warning when another property is a sibling of a $ref", () => {
-      const spec = {
-        paths: {
-          "/CoolPath": {
-            schema: {
-              $ref: "#/definitions/abc",
-              description: "My very cool schema"
-            }
-          }
-        },
-        definitions: {
-          abc: {}
-        }
-      }
-
-      return validateHelper(spec)
-      .then(system => {
-        const allErrors = system.errSelectors.allErrors().toJS()
-        expect(allErrors.length).toEqual(1)
-        const firstError = allErrors[0]
-        expect(firstError.message).toMatch("Sibling values are not allowed alongside $refs")
-        expect(firstError.level).toEqual("warning")
-        expect(firstError.path).toEqual(["paths", "/CoolPath", "schema", "description"])
-      })
-    })
-
-  })
 
   describe("Unused definitions", () => {
